+++ changelog.md
@@ -1,1 +1,1 @@
-- addedmore parameters to the recipes allowing to blacklist woods and select which wood per biome
+- fixed saved sawmill recipes not getting read on second reload